Load each stop_tokens token group into a Tensor
Evaluation of stop_tokens fails when is a jagged array, because a jagged array cannot be converted to a Tensor. Thus all token groups must be the same length. This precludes stop_tokens being passed as, for example, `[[187,187],[535]]` which are both representations of `\n\n`.

A _jagged array_ is an array of arrays, a vector of vectors or a list of lists where each member of the outer list is another list which may be of a different length to any or all of its siblings. Another example of a jagged array is `[[1], [2,3], [4,5,6]]`