fix rope precision bug
When we use model.bfloat16().cuda(), `inv_freq` will be converted to bfloat16 even we defined it in dtype=float32. In bfloat16, position would not
be represented right, for example, we cannot distinguish the difference between `256.0` and `257.0`, so when we create
position embedding, we forced convert `inv_freq` and `t` to fp32.